+++ LMSAdmin/src/main/java/com/lms/LMSAdmin/controller/LibraryBranchController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -14,49 +15,80 @@ import com.lms.LMSAdmin.pojo.LibraryBranch;
 import com.lms.LMSAdmin.service.LibraryBranchService;
 
 @RestController
+@RequestMapping("/LMSAdmin/libraryBranch")
 public class LibraryBranchController {
 	
 	@Autowired
 	LibraryBranchService branService;
 	
 	//Create a record
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/1/branchName/{name}/branchAddress/{address}", method = {RequestMethod.POST, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.CREATED, reason = "Library Branch record created.")
-	public void insertBranch(@PathVariable("name") String branchName, @PathVariable("address") String branchAddress) {
+	@RequestMapping(value = "/create/branchName/{name}/branchAddress/{address}", 
+			method = {RequestMethod.POST, RequestMethod.GET})
+	
+	public ResponseEntity<String> insertBranch(@PathVariable("name") String branchName, @PathVariable("address") String branchAddress) {
+		
 		branService.insertBranch(branchName, branchAddress);
+		return new ResponseEntity<String>("Library Branch record created.", HttpStatus.CREATED);
 	}
 	
 	//Update a record
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/2/branchId/{branchId}/branchName/{name}/branchAddress/{address}", method = {RequestMethod.PUT, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.OK, reason = "Library Branch record updated.")
-	public void updateBranch(@PathVariable("branchId") int branchId, @PathVariable("name") String branchName, @PathVariable("address") String branchAddress) {
-		branService.updateBranch(branchId, branchName, branchAddress);
+	@RequestMapping(value = "/update/branchId/{branchId}/branchName/{name}/branchAddress/{address}", 
+			method = {RequestMethod.PUT, RequestMethod.GET})
+	
+	public ResponseEntity<String> updateBranch(@PathVariable("branchId") int branchId, @PathVariable("name") String branchName, 
+			@PathVariable("address") String branchAddress) {
+		
+		boolean checkId = branService.ifExists(branchId);
+		
+		if(checkId == true) {
+			branService.updateBranch(branchId, branchName, branchAddress);
+			return new ResponseEntity<String>("Library Branch record updated.", HttpStatus.ACCEPTED);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//View all available branches to dispatch books to
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/3/branchId/{branchId}/deleteOpt/1/newBranchId", method = RequestMethod.GET)
+	@RequestMapping(value = "/dispatchDelete/branchId/{branchId}/newBranchId", method = RequestMethod.GET)
 	@ResponseStatus(code = HttpStatus.OK)
 	public List<LibraryBranch> getAvailableBranch(@PathVariable("branchId") int branchId) {
 		return branService.getAvailableBranch(branchId);
 	}
 	
 	//Dispatch books then delete branch
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/3/branchId/{branchId}/deleteOpt/1/newBranchId/{newBranId}", method = {RequestMethod.DELETE, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "Library Branch record deleted.")
-	public void dispatchDeleteBranch(@PathVariable("branchId") int branchId, @PathVariable("newBranId") int newBranId) {
-		branService.dispatchBooks(branchId, newBranId);
-		branService.deleteBranch(branchId);
+	@RequestMapping(value = "/dispatchDelete/branchId/{branchId}//newBranchId/{newBranId}", 
+			method = {RequestMethod.DELETE, RequestMethod.GET})
+	
+	public ResponseEntity<String> dispatchDeleteBranch(@PathVariable("branchId") int branchId, @PathVariable("newBranId") int newBranId) {
+		
+		boolean checkId = branService.ifExists(branchId);
+		
+		if(checkId == true) {
+			branService.dispatchBooks(branchId, newBranId);
+			branService.deleteBranch(branchId);
+			return new ResponseEntity<String>(HttpStatus.NO_CONTENT);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//Delete branch
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/3/branchId/{branchId}/deleteOpt/2", method = {RequestMethod.PUT, RequestMethod.DELETE, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "Library Branch record deleted.")
-	public void deleteBranch(@PathVariable("branchId") int branchId) {
-		branService.deleteBranch(branchId);
+	@RequestMapping(value = "/delete/libBranMenu/3/branchId/{branchId}", 
+			method = {RequestMethod.PUT, RequestMethod.DELETE, RequestMethod.GET})
+	
+	public ResponseEntity<String> deleteBranch(@PathVariable("branchId") int branchId) {
+		boolean checkId = branService.ifExists(branchId);
+		
+		if(checkId == true) {
+			branService.deleteBranch(branchId);
+			return new ResponseEntity<String>(HttpStatus.NO_CONTENT);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//View all records
-	@RequestMapping(value = "/lmsAdmin/menu/4/libBranMenu/4", method = RequestMethod.GET)
+	@RequestMapping(value = "/view", method = RequestMethod.GET)
 	@ResponseStatus(code = HttpStatus.OK)
 	public List<LibraryBranch> getAllBranches() {
 		return branService.getAllBranches();
