+++ LMSAdmin/src/main/java/com/lms/LMSAdmin/controller/PublisherController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -14,34 +15,56 @@ import com.lms.LMSAdmin.pojo.Publisher;
 import com.lms.LMSAdmin.service.PublisherService;
 
 @RestController
+@RequestMapping("/lmsAdmin/publisher")
 public class PublisherController {
 
 	@Autowired
 	PublisherService pubService;
 	
 	//Create a record
-	@RequestMapping(value = "/lmsAdmin/menu/3/pubMenu/1/pubName/{publisherName}/pubAddress/{publisherAddress}/pubPhone/{publisherPhone}", method = {RequestMethod.POST, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.CREATED, reason = "Publisher record created.")
-	public void insertPub(@PathVariable("publisherName") String pubName, @PathVariable("publisherAddress") String pubAddress, @PathVariable("publisherPhone") String pubPhone) {
+	@RequestMapping(value = "/create/pubName/{publisherName}/pubAddress/{publisherAddress}/pubPhone/{publisherPhone}", 
+			method = {RequestMethod.POST, RequestMethod.GET})
+	
+	public ResponseEntity<String> insertPub(@PathVariable("publisherName") String pubName, 
+			@PathVariable("publisherAddress") String pubAddress, @PathVariable("publisherPhone") String pubPhone) {
+		
 		pubService.insertPub(pubName, pubAddress, pubPhone);
+		return new ResponseEntity<String>("Publisher record created.", HttpStatus.CREATED);
 	}
 	
 	//Update a record
-	@RequestMapping(value = "/lmsAdmin/menu/3/pubMenu/2/pubId/{publisherId}/pubName/{publisherName}/pubAddress/{publisherAddress}/pubPhone/{publisherPhone}", method = {RequestMethod.PUT, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.OK, reason = "Publisher record updated.")
-	public void updatePub(@PathVariable("publisherId") int pubId, @PathVariable("publisherName") String pubName, @PathVariable("publisherAddress") String pubAddress, @PathVariable("publisherPhone") String pubPhone) {
-		pubService.updatePub(pubId, pubName, pubAddress, pubPhone);
+	@RequestMapping(value = "/update/pubId/{publisherId}/pubName/{publisherName}/pubAddress/{publisherAddress}/pubPhone/{publisherPhone}", 
+			method = {RequestMethod.PUT, RequestMethod.GET})
+	
+	public ResponseEntity<String> updatePub(@PathVariable("publisherId") int pubId, @PathVariable("publisherName") String pubName, 
+			@PathVariable("publisherAddress") String pubAddress, @PathVariable("publisherPhone") String pubPhone) {
+		
+		boolean checkId = pubService.ifExists(pubId);
+		
+		if(checkId == true) {
+			pubService.updatePub(pubId, pubName, pubAddress, pubPhone);
+			return new ResponseEntity<String>("Publisher record updated.", HttpStatus.ACCEPTED);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//Delete a record
-	@RequestMapping(value = "/lmsAdmin/menu/3/pubMenu/3/pubId/{publisherId}", method = {RequestMethod.DELETE, RequestMethod.GET})
-	@ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "Publisher record deleted.")
-	public void deletePub(@PathVariable("publisherId") int pubId) {
-		pubService.deletePub(pubId);
+	@RequestMapping(value = "/delete/pubId/{publisherId}", method = {RequestMethod.DELETE, RequestMethod.GET})
+	public ResponseEntity<String> deletePub(@PathVariable("publisherId") int pubId) {
+		
+		boolean checkId = pubService.ifExists(pubId);
+		
+		if(checkId == true) {
+			pubService.deletePub(pubId);
+			return new ResponseEntity<String>(HttpStatus.NO_CONTENT);
+		}else {
+			return new ResponseEntity<String>("Invalid ID.", HttpStatus.NOT_FOUND);
+		}
 	}
 	
 	//View all records
-	@RequestMapping(value = "/lmsAdmin/menu/3/pubMenu/4", method = RequestMethod.GET)
+	@RequestMapping(value = "/view", method = RequestMethod.GET)
 	@ResponseStatus(code = HttpStatus.OK)
 	public List<Publisher> getAllPubs() {
 		return pubService.getAllPubs();
